1.  linebreaker:  bLineOnScreen seems to be getting set incorrectly.  I added a comment to gripe about this, and fixed the symptom, which was that things were not being erased properly, resulting in screen dirt.
2.  fixed a couple of compiler warnings.

3.  In blocklayout, renamed format() to complete_format() and renamed
reformat() to minor_reformat() in order to more clearly show their purpose.
Note, however, that complete_format() is NOT currently a complete clobber
and reformat, as it was originally intended to be.  It preserves the runs,
but it removes the lines.  minor_reformat(), on the other hand, preserves
the runs AND the lines.  Neither one of these clobbers the slices, which
is currently a problem.

4.  fixed minor_reformat() to get rid of empty lines.  This entailed the
addition of Line::isEmpty().

5.  fixed the code which handles deleting things to use minor_reformat()
instead of complete_format().  Less flicker now.

6.  Fixed _repositionSlices() to call setNeedsCompleteReformat() for the
first block of the next column.

All of this was motivated by an attempt to fix the following bug:

Deleting things can cause empty space to appear at the bottom of a column,
but blocks never get slurped from the next column to fill in the newly
available space.

However, the motivating bug remains unfixed for the moment, since
there is currently no way to clobber the slices of a block and
reformat.  Attempts to introduce a way of doing so became hairy, so
I'm checking in what I've got now before I dive further.