Fixed bug #50 and #34 having to do with asserts in/around undo/redo and dealing with coalescing of fragments.
Fixed problems with Test_Dump counter.

Added assert to ap_Frame::getTitle() to assert if the title is longer
than the buffer the caller passed -- we need to fix this function to
actually respect the argument.

Added memset()'s to m_stTitle and m_szNondecoratedTitleName to keep
from getting garbage in calls to ap_Frame::getTitle().