A set of hacks which cause spell checking to more-or-less work. The primary hacks are:
1.  Every time a block needs some portion of itself to be spell
checked, we spell check the entire block.  This is inefficient,
but as a short-term workaround, it saves us from having to manage
a list of not-spell-checked regions of the block.

2.  Just after a block is spell checked, it's necessary to force
a redraw so that its squiggles will show up.  This is a tricky case,
which we will want to address specially when we do a major revisiting
of our redraw code.  For now, I am able to get the proper behavior by
erasing the insertion point, redrawing the entire block, and then
redrawing the insertion point.  This yields more screen activity than
we need, but as I said, I think we should fix that problem in the
context of an effort to fix several other similar ones.