bug fix.  the bug, which was never logged, was:
I opened Interview.abw and deleted from the middle of one paragraph to
the middle of another paragraph, including several paragraphs between
them.  Then, I hit undo.  The result caused an assert.  The assert appears
to have been due to problems which occur when you create new lines in
a column which needs a layout update.