1.  I think I fixed Bob's "bug from hell".  It was an uninitialized variable.
2.  Lots more image handling code has been added.  On Win32, you should
now be able to paste an image from the clipboard, save the resulting file (which
will have the PNG included), and load the file back in.

3.  Obviously, on Win32, you now need libpng and zlib to be able to build.

4.  All the image code, for now, is VERY constrained to a 24-bit format.
For many situations, this is abusive of memory, so I think we'll need to
eventually add support for other bit depths.

5.  ByteBuf can now write its contents to a file.  I needed this for testing,
and we'll need it later anyway.

6.  I added the concept of an imageFactory (see Design Patterns for the Factory
concept) at AP_App.

7.  Given the amount of code I touched in UNIX-specific stuff, I seriously
doubt this will build on Linux.  However, I'll get it working ASAP.