Massive set of changes to redesign how we handle span formatting hints.  TempSpanFmt and PreferredSpanFmt are now gone.  In their place is FmtMark.  This fixes or partially fixes or affects bugs 33, 218, 241, 242, 243, and 260.
There are still some minor quirks with it and TODO's, which I'll
fix up or log shortly -- but I needed to get this checked in today
so I can switch machines more easily (if only NT had decent network
support (can you say telnetd)).

A full clean build is required.

[To the Debian folks: I wait a few days before picking this up
and putting it in the distribution.]