Enabled all of the Non-latin-1 keybindings that i #if 0'd back when i added all of the non-us-ascii keybindings.
NOTE THIS IS STILL BROKEN ON WIN95/98, since they are based
upon code pages rather than unicode.  a keyevent arrives as
a code-page-relative value (where all the non-latin-1 chars
are folded into 0x80-0xff in the current code page).  i'd
like to be able to use ToUnicode(), but that doesn't exist
on 95/98.  I haven't tested it on NT yet.