Beginnings of a MacOS X build from Hubert Figuiere:
Here is the firsts patches for the MacOS X build. This is aimed to help
MacOS port to go forward as it require less work that a rough MacOS port
(without POSIX and other UNIXies).

The plan is to build MacOS X version using Carbon, then build the MacOS
version using Classic MacOS API (from which Carbon is derived) with
CodeWarrior and/or MPW.