Resurect support for libxml. This was known as Gnome-XML support and was enabled with OPT_ABI_GNOME_XML2, but not anymore; this commit makes it known as LibXML support (since libxml2 is not a gnome library, but rather one used by gnome), and makes the option for using it OPT_ABI_LIBXML2. More significantly, it lets AbiWord compile again using libxml. This is a large commit mostly because of:
	find . \( -name \*.c -o -name \*.cpp -o -name \*.h \) -print \
	| while read file ; do mv $file $file.XXX ; \
	sed -e 's/GNOME_XML2/LIBXML2/' < $file.XXX > $file ; rm $file.XXX ; \
	done

I am aware that a new release is happening soon, and was reluctant to
commit such a relatively large change, but realized that compiling and
possibly having some bugs when using libxml2 is much better than it not
compiling at all with the libxml option, which was the case before this
commit. Please note that very few of the files listed below were actually
modified by a single byte, just the above shellcode I used caused the
modification time on the files to change, in turn causing CVS to be
confused.