From martin.
This patch enables rtf export past fields. It fixes Ben's bug when typing
over a fully selected list. It adds robustness to multi-view situations.
It enables import of footer.abw but Abi crashes if you try to exit after
importing it. This is a marginal improvement on crashing upon trying to
import footer.abw.