Fixes up various things in the maintenance of ispell's ignored words list and also implements persisting the "Ignore All" list to the document file. The "ignore all" word list now acts like a persistent, per-document custom dictionary.
All *alloc()/free() sets in ignored word maintenance changed to
new/delete.  Also eliminated the need for memory allocation for
scratch buffers for most common cases.  Ignored word maintenance isn't
exactly performance critical, but every little bit helps.

The signature of enumIgnores() was changed because the signature it
had would have been ineffective.  (enumIgnores() was not previously
called, but is called now for a UT_DEBUGMSG() in the AD_Document
destructor.)

Fixed a sort of interesting problem about duplicate entries.  When you
"ignore all" on some word, spell-check queues all the blocks in a
document for re-spell-checking.  That's done in the background.  For a
long document, it's easy to do "ignore all" on another instance of the
same word before the squiggle is removed.  That leads to an attempt at
adding a duplicate entry to the hash table, and that leads to an
assert.  The assert looks harmless to me, but since hash tables are
used all over the place, I didn't remove it.  Instead, I made ignored
word maintenance check for an entry before trying to add it.

Added two preference variables, SpellCheckIgnoredWordsSave and
SpellCheckIgnoredWordsLoad (sorry, no GUI for any platform).  As a
side effect of the implementation to use those preferences,
PD_Document and AP_Convert objects now have an XAP_App pointer (and a
getApp()).  The means that lots of new places (including the
import/export code) can now access the user preferences.  Go hog wild.

Ignored words are saved in a bit of XML that looks like this:

	<ignoredwords>
	  <iw>sumthing</iw>
	  <iw>utherthing</iw>
	</ignoredwords>

The <ignoredwords> tag is at the same level as <section> and <style>.

Most changes are XP, tested on Linux.  I'm not sure if this works with
the alternative Gnome XML parser, but my guess is that it does.  A
small part of the changes is platform-specific, but things should work
if the code meets my modest assumptions about class hierarchies and
member functions.  The Unix stuff works for sure.