Updated pt-PT locale.
* SPECIAL NOTE: *

In

src/wp/ap/unix/ap_UnixApp.cpp

from lines 252 and on, abi seems to load improperly the string sets.

These are the differences:

            if (szDirectory[szPathname.size()-1]!='/') // was: ==
              szPathname += "/";
            szPathname += szStringSet;
            szPathname += ".strings"; // was nothing

Previous to this patch (with assistance from Dom) it tried to load
/usr/share/abisuite/AbiWord/stringspt-PT

and then failed.

Now, it loads

/usr/share/abisuite/AbiWord/strings/pt-PT.strings flawlessly.

This shouldn't break anything, but please, feel free to slap me with a long dead fish if it does.

Hugs, rms

PS: first code patch ;) thanks Dom.