(Frodo) Psion importer updates, Psion exporter major updates
YOU NEED PSICONV-0.9.4 TO COMPILE THIS PLUGIN.

This commit (HEAD only) solves a few buglets in the Psion importer.
It is contains an almost-rewrite of the exporter. The exporter is not
finished yet, but it should already work better than before this commit.

Highlights (exporter):
  * Cleanup of the class hierarchy
  * Cleanup of the source code
  * Updated comments
  * Support for Styles
  * Much improved handling of non-ASCII characters

Highlights (importer):
  * Cleanup of the included headers
  * A few bugs solved.

Psiconv 0.9.4 has been released today, and can be found at
http://huizen.dds.nl/~frodol/psiconv/ or http://debian.frodo.looijaard.name/
The Psiconv tree hosted with AbiWord is based on the 0.8.x tree and can not
be used! The configure fragment will detect whether your Psiconv version is
up-to-date.

Note that we are still not quite ready for a backport to the stable
tree (ignoring for the moment whether that would be desirable at all).