Start of work to get the RTF exporter to use unicode escaping for all PCData fields. This is necessary for representing chinese characters in such field. So far, this adds such functionality to style names and document properties (we already have it for font names).
Should be backported if possible.

Note of interest: Microsoft Word appears to lack such functionality. Word
sets the codepage of the RTF file based on the user's locale, as we do.
Word gives each font it's own code page, which should get around encoding
problems in the document itself. However, it I type Korean characters into the
document properties dialog of Word on a Chinese edition of XP. When I export to
RTF these characters get changed to "?".