Fix to #9638
This changes the windows "drop down list" combos in the toolbar
to "drop down" combos. It then sets the child edit control to read only.
This simulates the previous behaviour, except that the combo behaves like a
GTK combo in that it can display items not in the list.

Its previous inability to do this meant that the toolbar would
not always update to reflect the formatting at the cursor.