build against hunspell (http://hunspell.sf.net/) instead of myspell.
hunspell will be replacing myspell in a future version of OpenOffice.org.
it is compatible with myspell's dictionaries and offers a lot of improvements
for non-western languages.

we can no longer build against a system version of myspell. we will always
build against our own copy of hunspell unless told otherwise.

this is bug 9820