Make collision handling work when globs are involved!!
Example: User A inserts an 'x' character in a table, while user B just deleted that table. If B was joining user A in this collaborations session, the delete on B's side will be undone, and the 'x' will be inserted. User A will will simply reject the 'delete table' GLOB coming from user B.