Ryan Pavlik SoC - Work toward Win32 plugin gui: store HINSTANCE/HMODULE in SessionManager
In order to open a dialog, we must tell Windows what image (exe/dll) contains it.  We must use an HINSTANCE to do so.  Plugins (with resources in their DLL) may get a HMODULE (==HINSTANCE) from ABI_PLUGIN_DECLARE, in theory.  Store this in the session manager, since it also keeps track of dialog id's.