Ryan Pavlik SoC - Modify secondary constructor for Win32DialogHelper
This is the one I created before - turns out that if we pass the hWnd (meaning we have our own runModal and s_dlgProc), we don't need to pass an XAP_Dialog.  Should make this more flexible and elegant.  (The next commit will fix AbiCollab, the only consumer of this constructor right now.