Ryan Pavlik SoC - More polishing of dialogs
Most major bugs are hammered out, and the rest of the major bugs are in Bugzilla awaiting a fresh day and an alert mind.  This is probably OK for 2.5.2 - without all the asserts of a debug build, there's really only one crasher I know of (trying to disconnect or delete an online account).