Ryan Pavlik - Restructure TCP Synchronizer
Before, we had a lot of code in the actual .h.  Now there is a .cpp, which we can probably even sort into just a big win32 chunk and a big unix chunk at someone else's discretion.  Hopefully this doesn't break the Unix build of the TCP backend - if it does, I'm sorry!