Possible fix for 11541 or something like it (abiword crashes when saving when no space left).
On my system, the open() and writes() succeeded, but no data was flushed 
until close(). This caused close() to return an error, which AbiWord 
ignored.

Before, AbiWord wouldn't report that it failed to save the file, 
because it thought it had. Now, at least it reports the failure.

I'm not sure if this fixes the reported bug, but it certainly fixes some 
strange behavior.