Export ODT to the clipboard. This use the new ie_exp_DocRangeListener class to create a new document containing just the selected range. This is then exported to ODT in a temporary file and then copied to the clipboard.
I have not yet added 
src/wp/impexp/xp/ie_exp_DocRangeListener.(h,cpp) to the windows MSVisual C++
project desription so the windows buildis broken until that is fixed. Also we 
will need to duplicate a lot of the code in src/wp/ap/gtk/*.h to the appropriate
files in the windows and osx builds to get this functionality.